pyproject.toml
[project]
name = "subway-eink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairocffi>=1.7.1",
    "cairosvg>=2.7.1",
    "gpiozero>=2.0.1; sys_platform == 'linux'",
    "numpy>=2.1.3",
    "nyct-gtfs>=2.0.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "IT8951 @ git+https://github.com/GregDMeyer/IT8951; sys_platform == 'linux'",
    "rpi-gpio>=0.7.1; sys_platform == 'linux'"
]
